feat: improves `createApp` form to modular RTK Query approach
- Replaces the monolithic create app component using Redux slices with a modular structure leveraging RTK Query for data fetching and mutation.
- Splits UI into focused, reusable components for app status, file upload, tags, and user groups
